+++ version.go
@@ -1,36 +1,37 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
+	"io"
+	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"runtime"
-	"runtime/debug"
+  "runtime/debug"
 	"strings"
-	"time"
 
-	"github.com/charmbracelet/lipgloss"
 	"github.com/charmbracelet/log"
 	"github.com/go-git/go-git/v5"
-	"github.com/ssotspace/gitspace/lib"
 )
+
+type ReleaseInfo struct {
+	TagName string `json:"tag_name"`
+	ID      int    `json:"id"`
+}
 
 var Version string
 
 func getCurrentVersion() (string, string) {
-	// Check if Version is set (injected during build)
 	if Version != "" {
 		return Version, ""
 	}
 
-	// Try to get the git commit hash
 	hash, err := getGitCommitHash()
 	if err == nil && hash != "" {
-		return hash[:7], hash // Return first 7 characters as version, full hash as commit
+		return hash[:7], hash
 	}
 
-	// If git commit hash is not available, try to get it from build info
 	if info, ok := debug.ReadBuildInfo(); ok {
 		for _, setting := range info.Settings {
 			if setting.Key == "vcs.revision" {
@@ -39,19 +40,16 @@ func getCurrentVersion() (string, string) {
 		}
 	}
 
-	// If all else fails, return "unknown"
 	return "unknown", ""
 }
 
 func getGitCommitHash() (string, error) {
-	// Try using git command
 	cmd := exec.Command("git", "rev-parse", "HEAD")
 	output, err := cmd.Output()
 	if err == nil {
 		return strings.TrimSpace(string(output)), nil
 	}
 
-	// If git command fails, try using go-git
 	repo, err := git.PlainOpen(".")
 	if err != nil {
 		return "", err
@@ -65,53 +63,16 @@ func getGitCommitHash() (string, error) {
 	return ref.Hash().String(), nil
 }
 
-func printVersionInfo(logger *log.Logger) {
-	titleStyle := lipgloss.NewStyle().Bold(true).Foreground(lipgloss.Color("#00FFFF"))
-	infoStyle := lipgloss.NewStyle().Foreground(lipgloss.Color("#FFFF00"))
-
-	// Local install information
-	localVersion, localCommit := getCurrentVersion()
-
-	fmt.Println(titleStyle.Render("\nLocal Install:"))
-	fmt.Printf("Version: %s\n", infoStyle.Render(localVersion))
-	if localCommit != "" {
-		fmt.Printf("Commit Hash: %s\n", infoStyle.Render(localCommit))
-	}
-
-	// Remote/latest version information
-	remoteRelease, err := lib.GetLatestGitHubRelease("ssotops", "gitspace")
-	if err != nil {
-		logger.Error("Error fetching remote version info", "error", err)
-		return
-	}
-
-	fmt.Println(titleStyle.Render("\nRemote/Latest Version:"))
-	fmt.Printf("Version: %s\n", infoStyle.Render(remoteRelease.TagName))
-	fmt.Printf("Released: %s\n", infoStyle.Render(remoteRelease.PublishedAt.Format(time.RFC3339)))
-
-	// Extract commit hash from release body if available
-	commitHash := lib.ExtractCommitHash(remoteRelease.Body)
-	if commitHash != "" {
-		fmt.Printf("Commit Hash: %s\n", infoStyle.Render(commitHash))
-	} else {
-		fmt.Println("Commit Hash: Not available")
-	}
-}
-
 func upgradeGitspace(logger *log.Logger) {
 	logger.Info("Upgrading Gitspace...")
 
-	// Define repository details
 	repo := "ssotops/gitspace"
 	binary := "gitspace"
 
-	// Determine OS and architecture
 	osName := runtime.GOOS
 	arch := runtime.GOARCH
 
-	// Fetch the latest release information
-	logger.Info("Fetching latest release information...")
-	releaseInfo, err := lib.GetLatestGitHubRelease("ssotops", "gitspace")
+	releaseInfo, err := fetchLatestReleaseInfo(repo)
 	if err != nil {
 		logger.Error("Failed to fetch latest release information", "error", err)
 		return
@@ -120,15 +81,12 @@ func upgradeGitspace(logger *log.Logger) {
 	version := releaseInfo.TagName
 	logger.Info("Latest version", "version", version)
 
-	// Construct the download URL for the specific asset
 	assetName := fmt.Sprintf("%s_%s_%s", binary, osName, arch)
 	if osName == "windows" {
 		assetName += ".exe"
 	}
 	downloadURL := fmt.Sprintf("https://github.com/%s/releases/download/%s/%s", repo, version, assetName)
 
-	// Download the binary
-	logger.Info("Downloading new version", "version", version, "os", osName, "arch", arch)
 	tempFile, err := downloadBinary(downloadURL)
 	if err != nil {
 		logger.Error("Failed to download binary", "error", err)
@@ -136,7 +94,6 @@ func upgradeGitspace(logger *log.Logger) {
 	}
 	defer os.Remove(tempFile)
 
-	// Make it executable (skip for Windows)
 	if osName != "windows" {
 		err = os.Chmod(tempFile, 0755)
 		if err != nil {
@@ -145,14 +102,12 @@ func upgradeGitspace(logger *log.Logger) {
 		}
 	}
 
-	// Get the path of the current executable
 	execPath, err := os.Executable()
 	if err != nil {
 		logger.Error("Failed to get current executable path", "error", err)
 		return
 	}
 
-	// Replace the current binary with the new one
 	err = os.Rename(tempFile, execPath)
 	if err != nil {
 		logger.Error("Failed to replace current binary", "error", err)
@@ -160,6 +115,28 @@ func upgradeGitspace(logger *log.Logger) {
 	}
 
 	logger.Info("Gitspace has been successfully upgraded!", "version", version)
+}
+
+func fetchLatestReleaseInfo(repo string) (*ReleaseInfo, error) {
+	url := fmt.Sprintf("https://api.github.com/repos/%s/releases/latest", repo)
+	resp, err := http.Get(url)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var releaseInfo ReleaseInfo
+	err = json.Unmarshal(body, &releaseInfo)
+	if err != nil {
+		return nil, err
+	}
+
+	return &releaseInfo, nil
 }
 
 func downloadBinary(url string) (string, error) {
@@ -181,4 +158,20 @@ func downloadBinary(url string) (string, error) {
 	}
 
 	return tempFile.Name(), nil
+}
+
+func printVersionInfo(logger *log.Logger) {
+    version, commitHash := getCurrentVersion()
+    logger.Info("Current version", "version", version)
+    if commitHash != "" {
+        logger.Info("Commit hash", "hash", commitHash)
+    }
+
+    releaseInfo, err := fetchLatestReleaseInfo("ssotops/gitspace")
+    if err != nil {
+        logger.Error("Failed to fetch latest release information", "error", err)
+        return
+    }
+
+    logger.Info("Latest version", "version", releaseInfo.TagName)
 }
